+++ examples/generic-types/paginated-response.type.ts
@@ -3,8 +3,7 @@ import { ClassType, Field, ObjectType, Int } from "type-graphql";
 export function PaginatedResponse<TItemsFieldValue>(
   itemsFieldValue: ClassType<TItemsFieldValue> | string | number | boolean,
 ) {
-  // 'isAbstract' decorator option is mandatory to prevent registering in schema
-  @ObjectType({ isAbstract: true })
+  @ObjectType()
   abstract class PaginatedResponseClass {
     @Field(_type => [itemsFieldValue])
     items: TItemsFieldValue[];