+++ examples/resolvers-inheritance/recipe/recipe.resolver.ts
@@ -1,5 +1,5 @@
+import { Service } from "typedi";
 import { Resolver, FieldResolver, Root } from "type-graphql";
-
 import { ResourceResolver } from "../resource/resource.resolver";
 import { Recipe } from "./recipe.type";
 
@@ -11,7 +11,8 @@ const recipes: Recipe[] = [
   },
 ];
 
-@Resolver(of => Recipe)
+@Resolver(_of => Recipe)
+@Service()
 export class RecipeResolver extends ResourceResolver(Recipe, recipes) {
   // here you can add resource-specific operations