+++ src/metadata/definitions/resolver-metadata.ts
@@ -50,6 +50,5 @@ export interface SubscriptionResolverMetadata extends ResolverMetadata {
 export interface ResolverClassMetadata {
   target: Function;
   getObjectType: ClassTypeResolver;
-  isAbstract?: boolean;
   superResolver?: ResolverClassMetadata;
 }