+++ src/schema/schema-generator.ts
@@ -58,20 +58,17 @@ import {
   getObjectTypeDefinitionNode,
 } from "./definition-node";
 
-interface AbstractInfo {
-  isAbstract: boolean;
-}
-interface ObjectTypeInfo extends AbstractInfo {
+interface ObjectTypeInfo {
   target: Function;
   type: GraphQLObjectType;
   metadata: ObjectClassMetadata;
 }
-interface InterfaceTypeInfo extends AbstractInfo {
+interface InterfaceTypeInfo {
   target: Function;
   type: GraphQLInterfaceType;
   metadata: InterfaceClassMetadata;
 }
-interface InputObjectTypeInfo extends AbstractInfo {
+interface InputObjectTypeInfo {
   target: Function;
   type: GraphQLInputObjectType;
 }
@@ -88,7 +85,7 @@ export interface SchemaGeneratorOptions extends BuildContextOptions {
   /**
    * Array of resolvers classes
    */
-  resolvers?: Function[];
+  resolvers: Function[];
   /**
    * Array of orphaned type classes that are not used explicitly in GraphQL types definitions
    */
@@ -133,7 +130,7 @@ export abstract class SchemaGenerator {
     getMetadataStorage().build(options);
     this.buildTypesInfo(options.resolvers);
 
-    const orphanedTypes = options.orphanedTypes || (options.resolvers ? [] : undefined);
+    const orphanedTypes = options.orphanedTypes ?? [];
     const prebuiltSchema = new GraphQLSchema({
       query: this.buildRootQueryType(options.resolvers),
       mutation: this.buildRootMutationType(options.resolvers),
@@ -189,7 +186,7 @@ export abstract class SchemaGenerator {
       : defaultValueFromInitializer;
   }
 
-  private static buildTypesInfo(resolvers?: Function[]) {
+  private static buildTypesInfo(resolvers: Function[]) {
     this.unionTypesInfo = getMetadataStorage().unions.map<UnionTypeInfo>(unionMetadata => {
       // use closure to capture values from this selected schema build
       const unionObjectTypesInfo: ObjectTypeInfo[] = [];
@@ -266,7 +263,6 @@ export abstract class SchemaGenerator {
       return {
         metadata: objectType,
         target: objectType.target,
-        isAbstract: objectType.isAbstract || false,
         type: new GraphQLObjectType({
           name: objectType.name,
           description: objectType.description,
@@ -314,17 +310,11 @@ export abstract class SchemaGenerator {
             let fields = fieldsMetadata.reduce<GraphQLFieldConfigMap<any, any>>(
               (fieldsMap, field) => {
                 const { fieldResolvers } = getMetadataStorage();
-                const filteredFieldResolversMetadata = !resolvers
-                  ? fieldResolvers
-                  : fieldResolvers.filter(
-                      it => it.kind === "internal" || resolvers.includes(it.target),
-                    );
+                const filteredFieldResolversMetadata = fieldResolvers.filter(
+                  it => it.kind === "internal" || resolvers.includes(it.target),
+                );
                 const fieldResolverMetadata = filteredFieldResolversMetadata.find(
-                  it =>
-                    it.getObjectType!() === field.target &&
-                    it.methodName === field.name &&
-                    (it.resolverClassMetadata === undefined ||
-                      it.resolverClassMetadata.isAbstract === false),
+                  it => it.getObjectType!() === field.target && it.methodName === field.name,
                 );
                 const type = this.getGraphQLOutputType(
                   field.target,
@@ -401,7 +391,6 @@ export abstract class SchemaGenerator {
         return {
           metadata: interfaceType,
           target: interfaceType.target,
-          isAbstract: interfaceType.isAbstract || false,
           type: new GraphQLInterfaceType({
             name: interfaceType.name,
             description: interfaceType.description,
@@ -441,9 +430,7 @@ export abstract class SchemaGenerator {
                   const fieldResolverMetadata = getMetadataStorage().fieldResolvers.find(
                     resolver =>
                       resolver.getObjectType!() === field.target &&
-                      resolver.methodName === field.name &&
-                      (resolver.resolverClassMetadata === undefined ||
-                        resolver.resolverClassMetadata.isAbstract === false),
+                      resolver.methodName === field.name,
                   );
                   const type = this.getGraphQLOutputType(
                     field.target,
@@ -509,7 +496,6 @@ export abstract class SchemaGenerator {
       const inputInstance = new (inputType.target as any)();
       return {
         target: inputType.target,
-        isAbstract: inputType.isAbstract || false,
         type: new GraphQLInputObjectType({
           name: inputType.name,
           description: inputType.description,
@@ -556,7 +542,7 @@ export abstract class SchemaGenerator {
     });
   }
 
-  private static buildRootQueryType(resolvers?: Function[]): GraphQLObjectType {
+  private static buildRootQueryType(resolvers: Function[]): GraphQLObjectType {
     const queriesHandlers = this.filterHandlersByResolvers(getMetadataStorage().queries, resolvers);
 
     return new GraphQLObjectType({
@@ -565,7 +551,7 @@ export abstract class SchemaGenerator {
     });
   }
 
-  private static buildRootMutationType(resolvers?: Function[]): GraphQLObjectType | undefined {
+  private static buildRootMutationType(resolvers: Function[]): GraphQLObjectType | undefined {
     const mutationsHandlers = this.filterHandlersByResolvers(
       getMetadataStorage().mutations,
       resolvers,
@@ -580,7 +566,7 @@ export abstract class SchemaGenerator {
     });
   }
 
-  private static buildRootSubscriptionType(resolvers?: Function[]): GraphQLObjectType | undefined {
+  private static buildRootSubscriptionType(resolvers: Function[]): GraphQLObjectType | undefined {
     const subscriptionsHandlers = this.filterHandlersByResolvers(
       getMetadataStorage().subscriptions,
       resolvers,
@@ -595,7 +581,7 @@ export abstract class SchemaGenerator {
     });
   }
 
-  private static buildOtherTypes(orphanedTypes?: Function[]): GraphQLNamedType[] {
+  private static buildOtherTypes(orphanedTypes: Function[]): GraphQLNamedType[] {
     const autoRegisteredObjectTypesInfo = this.objectTypesInfo.filter(typeInfo =>
       typeInfo.metadata.interfaceClasses?.some(interfaceClass => {
         const implementedInterfaceInfo = this.interfaceTypesInfo.find(
@@ -614,18 +600,9 @@ export abstract class SchemaGenerator {
       }),
     );
     return [
-      ...this.filterTypesInfoByIsAbstractAndOrphanedTypesAndExtractType(
-        this.objectTypesInfo,
-        orphanedTypes,
-      ),
-      ...this.filterTypesInfoByIsAbstractAndOrphanedTypesAndExtractType(
-        this.interfaceTypesInfo,
-        orphanedTypes,
-      ),
-      ...this.filterTypesInfoByIsAbstractAndOrphanedTypesAndExtractType(
-        this.inputTypesInfo,
-        orphanedTypes,
-      ),
+      ...this.filterTypesInfoByOrphanedTypesAndExtractType(this.objectTypesInfo, orphanedTypes),
+      ...this.filterTypesInfoByOrphanedTypesAndExtractType(this.interfaceTypesInfo, orphanedTypes),
+      ...this.filterTypesInfoByOrphanedTypesAndExtractType(this.inputTypesInfo, orphanedTypes),
       ...autoRegisteredObjectTypesInfo.map(typeInfo => typeInfo.type),
     ];
   }
@@ -634,10 +611,6 @@ export abstract class SchemaGenerator {
     handlers: ResolverMetadata[],
   ): GraphQLFieldConfigMap<T, U> {
     return handlers.reduce<GraphQLFieldConfigMap<T, U>>((fields, handler) => {
-      // omit emitting abstract resolver fields
-      if (handler.resolverClassMetadata && handler.resolverClassMetadata.isAbstract) {
-        return fields;
-      }
       const type = this.getGraphQLOutputType(
         handler.target,
         handler.methodName,
@@ -666,11 +639,6 @@ export abstract class SchemaGenerator {
     const { pubSub, container } = BuildContext;
     const basicFields = this.generateHandlerFields(subscriptionsHandlers);
     return subscriptionsHandlers.reduce<GraphQLFieldConfigMap<T, U>>((fields, handler) => {
-      // omit emitting abstract resolver fields
-      if (handler.resolverClassMetadata && handler.resolverClassMetadata.isAbstract) {
-        return fields;
-      }
-
       let subscribeFn: GraphQLFieldResolver<T, U>;
       if (handler.subscribe) {
         subscribeFn = handler.subscribe;
@@ -871,17 +839,15 @@ export abstract class SchemaGenerator {
 
   private static filterHandlersByResolvers<T extends ResolverMetadata>(
     handlers: T[],
-    resolvers: Function[] | undefined,
+    resolvers: Function[],
   ) {
-    return resolvers ? handlers.filter(query => resolvers.includes(query.target)) : handlers;
+    return handlers.filter(query => resolvers.includes(query.target));
   }
 
-  private static filterTypesInfoByIsAbstractAndOrphanedTypesAndExtractType(
+  private static filterTypesInfoByOrphanedTypesAndExtractType(
     typesInfo: Array<ObjectTypeInfo | InterfaceTypeInfo | InputObjectTypeInfo>,
-    orphanedTypes: Function[] | undefined,
+    orphanedTypes: Function[],
   ) {
-    return typesInfo
-      .filter(it => !it.isAbstract && (!orphanedTypes || orphanedTypes.includes(it.target)))
-      .map(it => it.type);
+    return typesInfo.filter(it => orphanedTypes.includes(it.target)).map(it => it.type);
   }
 }