+++ tests/functional/generic-types.ts
@@ -30,8 +30,8 @@ describe("Generic types", () => {
     getMetadataStorage().clear();
   });
 
-  it("shouldn't emit abstract object type", async () => {
-    @ObjectType({ isAbstract: true })
+  it("shouldn't emit unused abstract object type", async () => {
+    @ObjectType()
     abstract class BaseType {
       @Field()
       baseField: string;
@@ -65,8 +65,8 @@ describe("Generic types", () => {
     expect(baseTypeInfo).toBeUndefined();
   });
 
-  it("shouldn't emit abstract interface type", async () => {
-    @InterfaceType({ isAbstract: true })
+  it("shouldn't emit unused abstract interface type", async () => {
+    @InterfaceType()
     abstract class BaseInterfaceType {
       @Field()
       baseField: string;
@@ -111,8 +111,8 @@ describe("Generic types", () => {
     expect(baseInterfaceTypeInfo).toBeUndefined();
   });
 
-  it("shouldn't emit abstract input object type", async () => {
-    @InputType({ isAbstract: true })
+  it("shouldn't emit unused abstract input object type", async () => {
+    @InputType()
     abstract class BaseInput {
       @Field()
       baseField: string;
@@ -152,7 +152,7 @@ describe("Generic types", () => {
 
     beforeEach(async () => {
       function Connection<TItem>(TItemClass: ClassType<TItem>) {
-        @ObjectType(`${TItemClass.name}Connection`, { isAbstract: true })
+        @ObjectType(`${TItemClass.name}Connection`)
         class ConnectionClass {
           @Field(() => Int)
           count: number;
@@ -264,7 +264,7 @@ describe("Generic types", () => {
 
     beforeEach(async () => {
       function Edge<TNodeClass>(NodeClass: ClassType<TNodeClass>) {
-        @ObjectType({ isAbstract: true })
+        @ObjectType()
         abstract class EdgeClass {
           @Field(() => NodeClass)
           node: TNodeClass;
@@ -410,7 +410,7 @@ describe("Generic types", () => {
 
     beforeAll(async () => {
       function Base<TType>(TTypeClass: ClassType<TType>) {
-        @ObjectType({ isAbstract: true })
+        @ObjectType()
         class BaseClass {
           @Field(() => TTypeClass)
           baseField: TType;