+++ tests/functional/query-complexity.ts
@@ -84,7 +84,7 @@ describe("Query complexity", () => {
       }
 
       function createResolver(name: string, objectType: ClassType) {
-        @Resolver(() => objectType, { isAbstract: true })
+        @Resolver(() => objectType)
         class BaseResolver {
           protected name = "baseName";